First round of imports from HEAD

+++ servers/slapd/back-ldap/config.c
@@ -1,7 +1,7 @@
 /* config.c - ldap backend configuration file routine */
 /* $OpenLDAP$ */
 /*
- * Copyright 1998-2002 The OpenLDAP Foundation, All Rights Reserved.
+ * Copyright 1998-2003 The OpenLDAP Foundation, All Rights Reserved.
  * COPYING RESTRICTIONS APPLY, see COPYRIGHT file
  */
 /* This is an altered version */
@@ -444,7 +444,7 @@ suffix_massage_config(
 			 */
 			return -1;
 
-		} else if ( len >= sizeof( vbuf_ ) ) {
+		} else if ( len >= (int)sizeof( vbuf_ ) ) {
 			/* 
 			 * C99: snprintf returns the required size 
 			 */
@@ -459,7 +459,7 @@ suffix_massage_config(
 		if ( len == -1 ) {
 			return -1;
 
-		} else if ( len >= sizeof( rbuf_ ) ) {
+		} else if ( len >= (int)sizeof( rbuf_ ) ) {
 			rbuf = ch_malloc( len + 1 );
 			len = snprintf( rbuf, sizeof( rbuf_ ), "%%1%s)%%2", 
 					nrnc->bv_val );
